Release v3.3.0
This release adds tolerance for errors - following was previously parseable:
{ "arr": [ 1, 2, 3 ] }

But this caused error
{ "arr": [ 1, 2, 3, ] }
just like this did
{ "obj": { "value1": 1, "value2": 2, }}

commas are supposed to tell that there is more on array or object, but
often it is that there is that extra comma without next element; this
is not valid json, but since there is no harm on this, I enabled this
tolerance while parsing json. Still; it tolerates examples shown, but
does not tolerate anything, for example:
{ "arr": [ 1, 2, 3,, ] }
will cause error; just one extra comma is tolerated, if it isn't enough,
json is too far away from valid json to be accepted.

I have updated sample file to demonstrate this tolerance.

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>