Release v4.2.5
This release gives improvements to validator.
Unfortunately, previous validator definitions are not completely compatible; allowed_values is now just values.

New feature with validator, is that it allows to embed child validators to object types, validator example
has been updated to include a sample.

Another feature is that there is a describe function to validation predicate, and also output stream
support is provided. Also this feature is shown in example.

Also one bug was fixed; foreach_iterator contained un-implemented key() function which was removed.
Correct function to use, is name().

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>